Убрано дублирование кода в клиентах RestAssured
