+++ pom.xml
@@ -46,7 +46,7 @@
 		<dependency>
 			<groupId>io.rest-assured</groupId>
 			<artifactId>rest-assured</artifactId>
-			<version>5.5.0</version>
+			<version>5.4.0</version>
 		</dependency>
 	</dependencies>
