+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -7,7 +7,6 @@ import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import lombok.Data;
 import ru.yandex.practicum.filmorate.validation.IsAfter;
-
 import java.time.LocalDate;
