+++ src/test/java/ru/yandex/practicum/filmorate/helper/BaseRestAssuredClient.java
@@ -2,6 +2,11 @@ package ru.yandex.practicum.filmorate.helper;
 
 import io.restassured.config.HttpClientConfig;
 import io.restassured.config.RestAssuredConfig;
+import io.restassured.response.Response;
+import static io.restassured.RestAssured.given;
+import static io.restassured.http.ContentType.JSON;
+import static org.junit.jupiter.api.Assertions.fail;
+
 
 public abstract class BaseRestAssuredClient {
 
@@ -14,5 +19,47 @@ public abstract class BaseRestAssuredClient {
                 .httpClient(HttpClientConfig.httpClientConfig()
                         .setParam("http.socket.timeout", 10000)
                         .setParam("http.connection.timeout", 10000));
+    }
+
+    public <T> Response sendPostRequest(T payload) {
+        Response response = null;
+        try {
+            response = given()
+                    .config(config)
+                    .contentType(JSON)
+                    .accept(JSON)
+                    .body(payload)
+                    .post(urlPrefix);
+        } catch (Exception e) {
+            fail("Exception occurred: " + e.getClass() + " " + e.getMessage());
+        }
+        return response;
+    }
+
+    public <T> Response sendPutRequest(T payload) {
+        Response response = null;
+        try {
+            response = given()
+                    .config(config)
+                    .contentType(JSON)
+                    .body(payload)
+                    .put(urlPrefix);
+        } catch (Exception e) {
+            fail("Exception occurred: " + e.getClass() + " " + e.getMessage());
+        }
+        return response;
+    }
+
+    public Response sendGetAllRequest() {
+        Response response = null;
+        try {
+            response = given()
+                    .config(config)
+                    .accept(JSON)
+                    .get(urlPrefix);
+        } catch (Exception e) {
+            fail("Exception occurred: " + e.getClass() + " " + e.getMessage());
+        }
+        return response;
     }
 }
